requirements.txt
urllib3
google-cloud-pubsub
google-cloud-bigquery
